# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramStarsInvoiceInput {
  /**
     * Telegram user ID used to associate the invoice payload with the current Mini App user.
     * @minimum 1
     * @nullable
     */
  telegramUserId?: number | null;
}

export interface TelegramStarsInvoice {
  invoiceLink: string;
  payload: string;
  amountStars: 1000;
  creditUsd: 10;
  currency: 'XTR';
}

export interface ApiError {
  error: string;
}