# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: telegram
    description: Telegram Stars deposit operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /telegram/stars-invoice:
    post:
      operationId: createTelegramStarsInvoice
      tags: [telegram]
      summary: Create a Telegram Stars deposit invoice
      description: Creates a real Telegram Stars invoice for exactly 1000 XTR, representing 10 USDT of advertising balance.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramStarsInvoiceInput"
      responses:
        "200":
          description: Telegram invoice link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramStarsInvoice"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Telegram invoice service is not configured or unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TelegramStarsInvoiceInput:
      type: object
      properties:
        telegramUserId:
          type: ["integer", "null"]
          minimum: 1
          description: Telegram user ID used to associate the invoice payload with the current Mini App user.
    TelegramStarsInvoice:
      type: object
      properties:
        invoiceLink:
          type: string
          format: uri
        payload:
          type: string
        amountStars:
          type: integer
          const: 1000
        creditUsd:
          type: number
          const: 10
        currency:
          type: string
          const: XTR
      required:
        - invoiceLink
        - payload
        - amountStars
        - creditUsd
        - currency
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error