# lib/api-zod/src/generated/types/telegramStarsInvoice.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramStarsInvoice {
  invoiceLink: string;
  payload: string;
  amountStars: 1000;
  creditUsd: 10;
  currency: 'XTR';
}