# lib/api-zod/src/generated/types/telegramStarsInvoiceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramStarsInvoiceInput {
  /**
     * Telegram user ID used to associate the invoice payload with the current Mini App user.
     * @minimum 1
     * @nullable
     */
  telegramUserId?: number | null;
}